# docs/Design/API/openapi.yaml
openapi: 3.0.3
info:
  title: Gym Management System API
  version: 1.0.2
  description: |
    OpenAPI 3.0 contract cho Module 1 Auth + Module 2 RBAC + Module 3 Package + Module 4 Member/Subscription/Payment + Module 6 Facility.
    Module 7 Training va Module 8 Feedback da detailed o Markdown spec, nhung OpenAPI paths pending sync sau doc review.
    Module 5 Staff va Module 9 Report defer session sau.
    Source-of-truth conventions: [docs/Design/API/conventions.md](./conventions.md).
    Source-of-truth error envelope: server/src/common/filters/http-exception.filter.ts.
  contact:
    name: Lê Thanh An
servers:
  - url: /api/v1
    description: Base URL (global prefix set in server/src/main.ts)
...
security:
  - BearerAuth: []
paths:
...
  /members/{id}/progress:
    parameters:
      - { $ref: '#/components/parameters/IdParam' }
    get:
      tags: [Members]
      summary: UC06 — List bản ghi tiến độ
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: sort, in: query, schema: { type: string, default: recorded_at:desc } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/MemberProgress' }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: BigInt PK serialized as string.
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
  schemas:
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
        message: { type: string }
    ErrorEnvelope:
      type: object
      required: [success, code, message]
      properties:
        success: { type: boolean, enum: [false] }
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        details:
          oneOf:
            - { type: array, items: { type: string } }
            - { type: object, additionalProperties: true }
    PaginationMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        totalItems: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    MemberProgress:
      type: object
      required: [progressId, memberId, staffId, recordedAt]
      properties:
        progressId: { type: string }
        memberId: { type: string }
        staffId: { type: string }
        staffName: { type: string }
        weight: { type: string, description: Decimal as string, example: '65.50' }
        bmi: { type: string, example: '22.10' }
        goal: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        recordedAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: JWT thiếu / sai / expired
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, code: UNAUTHORIZED, message: Token không hợp lệ }
    Forbidden:
      description: RolesGuard / OwnershipGuard reject
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, code: FORBIDDEN, message: Không có quyền truy cập }
    NotFound:
      description: Resource không tồn tại / soft-deleted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, code: NOT_FOUND, message: Không tìm thấy bản ghi yêu cầu }
...